requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
psycopg2==2.9.10
python-decouple==3.8
gunicorn==23.0.0
django-heroku==0.3.1
dj-database-url==2.3.0
whitenoise==6.9.0